[*] Windows comliler warnings fix

+++ modules/channel/chain/dap_stream_ch_chain.c
@@ -695,7 +695,7 @@ static bool s_gdb_in_pkt_proc_callback(dap_proc_thread_t *a_thread, void *a_arg)
 
         uint64_t l_last_id = l_store_obj->id;
         const char *l_last_group = l_store_obj->group;
-        int l_last_type = l_store_obj->type;
+        uint32_t l_last_type = l_store_obj->type;
         bool l_group_changed = false;
         uint32_t l_time_store_lim = dap_config_get_item_uint32_default(g_config, "resources", "dap_global_db_time_store_limit", 72);
         uint64_t l_limit_time = l_time_store_lim ? (uint64_t)time(NULL) - l_time_store_lim * 3600 : 0;
@@ -750,8 +750,8 @@ static bool s_gdb_in_pkt_proc_callback(dap_proc_thread_t *a_thread, void *a_arg)
             }
             time_t l_timestamp_del = global_db_gr_del_get_timestamp(l_obj->group, l_obj->key);
             // check the applied object newer that we have stored or erased
-            if (l_obj->timestamp >= l_timestamp_del &&
-                    l_obj->timestamp >= l_timestamp_cur &&
+            if (l_obj->timestamp >= (uint64_t)l_timestamp_del &&
+                    l_obj->timestamp >= (uint64_t)l_timestamp_cur &&
                     (l_obj->type != 'd' || l_obj->timestamp > l_limit_time)) {
                 l_apply = true;
             }
@@ -764,9 +764,9 @@ static bool s_gdb_in_pkt_proc_callback(dap_proc_thread_t *a_thread, void *a_arg)
                         l_store_obj[i].key, l_ts_str, l_store_obj[i].value_len);
             }
             if (!l_apply) {
-                if (l_obj->timestamp < l_timestamp_cur)
+                if (l_obj->timestamp < (uint64_t)l_timestamp_cur)
                     log_it(L_WARNING, "New data not applied, because newly object exists");
-                if (l_obj->timestamp < l_timestamp_del)
+                if (l_obj->timestamp < (uint64_t)l_timestamp_del)
                     log_it(L_WARNING, "New data not applied, because newly object is deleted");
                 if ((l_obj->type == 'd' && l_obj->timestamp <= l_limit_time))
                     log_it(L_WARNING, "New data not applied, because object is too old");
